Fixing broken `&` functionality when spaceless.nvim enabled
There's a built-in Vim/Neovim functionality of repeating last
substitution in current line with `&` key.

Spaceless.nvim has this functionality broken because of using `:s`
command internally to strip trailing whitespace, which redefines last
replacement command, making `&` non-functional (replaces search match
with blank space).

This PR is a quick demo fix for this, it uses Neovim API internally to
replace buffer line ranges. I cannot test in extensively now, probably
setting buffer to current (0) is a bad idea. Anyway, this is a base for
original plugin author to work with.